Hot-Fix for read file problems
* The Reader class is no longer extends Thread
* Moved code to readFileLines()

These changes would take effect temoprary while a "full" patch for this
problem is in progress. Due to performance impact and user experience,
it has it own branch and is strong recommended to be apllied.

Problem:
Cause of the async work of the reader the calling methods are getting
null pointer, they are causing NullPointerExceptions and may corrupt the
files and invoke bad use experience.